refactor(project-pattern): changing folders
