+++ .eslintrc.json
@@ -18,19 +18,6 @@
         "argsIgnorePattern": "^_"
       }
     ],
-    "react/prefer-stateless-function": "error",
-    "react/no-unused-prop-types": "error",
-    "react/jsx-pascal-case": "error",
-    "react/jsx-no-script-url": "error",
-    "react/no-danger": "error",
-    "react/no-danger-with-children": "error",
-    "react/no-unstable-nested-components": [
-      "error",
-      {
-        "allowAsProps": true
-      }
-    ],
-    "react/jsx-fragments": "error",
     "react/destructuring-assignment": [
       "error",
       "always",
@@ -58,24 +45,10 @@
         "namedComponents": "function-declaration"
       }
     ],
-    "react/jsx-key": [
-      "error",
-      {
-        "checkFragmentShorthand": true,
-        "checkKeyMustBeforeSpread": true,
-        "warnOnDuplicates": true
-      }
-    ],
-    "react/jsx-no-useless-fragment": "warn",
-    "react/jsx-curly-brace-presence": "warn",
-    "react/no-typos": "warn",
-    "react/display-name": "warn",
     "react/self-closing-comp": "warn",
     "react/jsx-sort-props": "warn",
     "react/react-in-jsx-scope": "off",
     "react/jsx-uses-react": "off",
-    "react/jsx-one-expression-per-line": "off",
-    "react/prop-types": "off",
     "quotes": [
       "error",
       "single"
@@ -84,19 +57,14 @@
       "error",
       "never"
     ],
-    "react/no-deprecated": "error",
     "react/button-has-type": "error",
     "react/jsx-boolean-value": 2,
     "react/jsx-closing-bracket-location": [
       1,
       "line-aligned"
     ],
-    "react/no-children-prop": "error",
     "react/jsx-props-no-spreading": [
       "warn"
-    ],
-    "react/no-unescaped-entities": [
-      "off"
     ],
     "no-var": "error",
     "no-unreachable": "error",
@@ -121,7 +89,7 @@
     "max-len": [
       "error",
       {
-        "code": 150
+        "code": 180
       }
     ],
     "jsx-quotes": [
