+++ src/components/events/EventContent.tsx
