+++ src/components/gallery/index.ts
