+++ src/components/hero/Detail.tsx
