+++ src/components/hero/styles/Detail.module.css
