+++ src/components/image_hero/ImageHero.tsx
@@ -5,15 +5,13 @@ import DevsNorteLight from '/public/icons/devsnorte-hero-light.svg'
 import Star from '/public/icons/star.svg'
 import { ThemeContext } from '../../contexts/ThemeContext'
 
-function ImageHero() {
+export function ImageHero() {
   const { theme } = useContext(ThemeContext)
   return (
-    <div className="relative">
+    <div className='relative'>
       {theme === 'dark' ? <DevsNorte className={styles.devsnorteSvg} /> : <DevsNorteLight className={styles.devsnorteSvg} />}
 
       <Star className={styles.star} />
     </div>
   )
 }
-
-export default ImageHero
