+++ src/components/image_hero/index.ts
