+++ src/components/newsletter/index.ts
