+++ src/components/section/Info.tsx
