+++ src/types/components/galleryTypes.ts
