+++ src/types/components/sectionTypes.ts
