--- src/types/gallery.ts
@@ -1,5 +1,0 @@
-export type ImageArray = string[];
-
-export interface GalleryProps {
-  images: ImageArray;
-}
